Plot2(IsoGeneGUI)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


Plot2(x, Myhscale, Myvscale, title)

Arguments

x 	 ~~Describe x here~~

Myhscale 	 ~~Describe Myhscale here~~

Myvscale 	 ~~Describe Myvscale here~~

title 	 ~~Describe title here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(x,Myhscale,Myvscale,title )
{

winPlot <- tktoplevel()
tkwm.title(winPlot ,title )
buttonFramePlot <-tkframe(winPlot ,borderwidth=2)

scr <- tkscrollbar(winPlot , repeatinterval=1, command=function(...)tkyview(tl,...))
tl <-tklistbox(winPlot ,height=2,selectmode="single",yscrollcommand=function(...)tkset(scr,...),background="white")

type <- c("ps","png","tiff","bmp","jpeg")

for (i in (1:length(type )))
{
    tkinsert(tl,"end",type [i])
  }
tkselection.set(tl,0)

img <- tkrplot(winPlot ,fun=x,hscale=Myhscale,vscale=Myvscale)

CopyToClip <- function()
{
  tkrreplot(img)
 tkmessageBox(message="The graph have been copied into clipboard",icon="info",type="ok")

  }

SaveGraph <- function()
{
fileName<- tclvalue(tkgetSaveFile())
if (!nchar(fileName))
    tkmessageBox(message="No graph to copy!")
else
{ 
  fileTypeChoice <- type[as.numeric(tkcurselection(tl))+1]
            split <- strsplit(fileName,".",fixed=TRUE)
            fileName <- paste(split[[1]][1],".",fileTypeChoice ,sep="")
            if(fileTypeChoice == "ps") fileTypeChoice <- "postscript"
            eval(parse(text=paste(fileTypeChoice ,"(file=fileName)",sep="")))
x()
            dev.off()
       }
  }

copy.but <- tkbutton(buttonFramePlot,text="Copy to Clipboard",command=CopyToClip)
save.but <- tkbutton(buttonFramePlot,text="Browse",command=SaveGraph)

tkgrid(img)
tkgrid(copy.but,tklabel(buttonFramePlot,text="Save image as:"),tl,scr,save.but)

tkgrid(buttonFramePlot)

  }

[Package IsoGeneGUI version 1.0 Index]