PvalAdj(IsoGeneGUI)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


PvalAdj(rp, FDR, type = c("BH", "BY", "Bonferroni", "Holm", "Hochberg", "SidakSS", "SidakSD"), stat = c("E2", "Williams", "Marcus", "M", "ModM"))

Arguments

rp 	 ~~Describe rp here~~

FDR 	 ~~Describe FDR here~~

type 	 ~~Describe type here~~

stat 	 ~~Describe stat here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (rp, FDR, type = c("BH", "BY","Bonferroni", "Holm", 
"Hochberg", "SidakSS", "SidakSD"), stat = c("E2", "Williams", 
    "Marcus", "M", "ModM")) 
{
    Probe.ID <- rp[, 1]
    type <- match.arg(type)
    stat <- match.arg(stat)
    if (stat == "E2") 
        rpraw <- rp[, 2]
    else if (stat == "Williams") 
        rpraw <- rp[, 3]
    else if (stat == "Marcus") 
        rpraw <- rp[, 4]
    else if (stat == "M") 
        rpraw <- rp[, 5]
    else if( stat == "ModM")
rpraw <- rp[, 6]
  
    adjp <- adjustment (rpraw , type ) 
colnames(adjp )<- c("raw p-val",paste(type,"p-val"))
   return(adjp )
  }

[Package IsoGeneGUI version 1.0 Index]