SEPlot(IsoGeneGUI)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


SEPlot(stat, xfudge)

Arguments

stat 	 ~~Describe stat here~~

xfudge 	 ~~Describe xfudge here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(stat,xfudge)
{
  stat1 <- get(paste(stat,"Val",sep="" ))
  se <- get(paste("Se",stat,"Val",sep="" ))
params <- par(bg="white")
par(mfrow=c(2,1))
xx <- cbind(se ,stat1 )
colors  <- densCols(xx)
plot(xx , col=colors,xlab="Standar error", ylab=stat)
      if(xfudge != 0) abline(v=xfudge,col="red" )
plot(density(SeModMVal ) ,xlab="Standar error",main="",lwd=2,col="blue")
if(xfudge != 0) abline(v=xfudge,col="red"  )
  }

[Package IsoGeneGUI version 1.0 Index]