Setseed(IsoGeneGUI)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


Setseed()

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function() {
seed <-tktoplevel()
tkwm.deiconify(seed)
tkwm.title(seed ,"Set random seed number") 
tkgrab.set(seed)
tkfocus(seed)
Name <- tclVar("1234")
entry.Name <-tkentry(seed ,width="10",textvariable=Name)
tkgrid(tklabel(seed ,text="Enter the random seed number"))
tkgrid(entry.Name)
OnOK <- function()
{
SeedNum <<- as.numeric(tclvalue(Name))
    assign("SeedNum",SeedNum, envir=.GlobalEnv)
tkdestroy(seed )
tkfocus(tt )
  }
OK.but <-tkbutton(seed ,text="   OK   ",command=OnOK)
tkbind(entry.Name, "<Return>",OnOK)
tkgrid(OK.but)
tkfocus(seed )
  }

[Package IsoGeneGUI version 1.0 Index]