xfudge(IsoGeneGUI)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


xfudge(x.dif, si2, inputFudge)

Arguments

x.dif 	 ~~Describe x.dif here~~

si2 	 ~~Describe si2 here~~

inputFudge 	 ~~Describe inputFudge here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--    or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(x.dif, si2, inputFudge ) {
  xs.alpha <- quantile(si2, (0:20) / 20)
  xs.perc <- quantile(si2, (0:100) / 100)
  xs.perc.gr <- as.numeric(cut(si2, breaks=xs.perc, include.lowest=TRUE))
  xs.mad <- matrix(0, length(xs.alpha), length(xs.perc)-1)
  for (ii in 1:100){
    k0 <- xs.perc.gr == ii
    xs.mad[,ii] <- apply(x.dif[k0]/outer(si2[k0],xs.alpha,"+"),2,mad, constant = 1)
  }
  xcv <- function(x) sqrt(var(x)) / mean(x)   # CV
  xs.cv <- apply(xs.mad, 1, xcv)
  if (inputFudge == "auto")  { xfudge <- xs.alpha[sort.list(xs.cv)[1]] }
  else   {
pct <- (0:20) *10/ 2
xfudge  <- xs.alpha[pct==inputFudge ]
  } 
  resFudge <- list ( xfudge,xs.cv )
  return(resFudge )
  }

[Package IsoGeneGUI version 1.0 Index]